requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
distro==1.9.0
fsspec==2024.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
json_repair==0.39.1
neo4j==5.28.1
neo4j-graphrag==1.7.0
lmstudio==1.3.1
pip==25.1.1
pydantic==2.11.5
pydantic_core==2.33.2
pypdf==5.5.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
tqdm==4.67.1
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
typing-inspection==0.4.1